Scan size component, raw, typo
